+++ little_brother/status_server.py
@@ -339,8 +339,7 @@ class StatusServer(base_web_server.BaseWebServer):
 
     def save_users_data(self, p_users, p_forms):
 
-        with little_brother.persistence.session_context.SessionContext(
-                p_persistence=self._persistence) as session_context:
+        with SessionContext(p_persistence=self._persistence) as session_context:
             session = session_context.get_session()
             changed = False
 
@@ -380,8 +379,7 @@ class StatusServer(base_web_server.BaseWebServer):
 
     def save_devices_data(self, p_devices, p_forms):
 
-        with little_brother.persistence.session_context.SessionContext(
-                p_persistence=self._persistence) as session_context:
+        with SessionContext(p_persistence=self._persistence) as session_context:
             session = session_context.get_session()
             changed = False
 
@@ -408,8 +406,7 @@ class StatusServer(base_web_server.BaseWebServer):
         with tools.TimingContext(lambda duration: self.measure(p_hostname=request.remote_addr,
                                                                p_service=self.simplify_url(request.url_rule),
                                                                p_duration=duration)):
-            with little_brother.persistence.session_context.SessionContext(
-                    p_persistence=self._persistence) as session_context:
+            with SessionContext(p_persistence=self._persistence) as session_context:
 
                 try:
                     admin_infos = self._appcontrol.get_admin_infos(p_session_context=session_context)
@@ -488,8 +485,7 @@ class StatusServer(base_web_server.BaseWebServer):
         with tools.TimingContext(lambda duration: self.measure(p_hostname=request.remote_addr,
                                                                p_service=self.simplify_url(request.url_rule),
                                                                p_duration=duration)):
-            with little_brother.persistence.session_context.SessionContext(
-                    p_persistence=self._persistence) as session_context:
+            with SessionContext(p_persistence=self._persistence) as session_context:
                 try:
                     topology_infos = self._appcontrol.get_topology_infos(p_session_context=session_context)
                     page = flask.render_template(
@@ -512,8 +508,7 @@ class StatusServer(base_web_server.BaseWebServer):
     @flask_login.login_required
     def users_view(self):
 
-        with little_brother.persistence.session_context.SessionContext(
-                p_persistence=self._persistence) as session_context:
+        with SessionContext(p_persistence=self._persistence) as session_context:
             request = flask.request
 
             with tools.TimingContext(lambda duration: self.measure(p_hostname=request.remote_addr,
@@ -616,8 +611,7 @@ class StatusServer(base_web_server.BaseWebServer):
     @flask_login.login_required
     def devices_view(self):
 
-        with little_brother.persistence.session_context.SessionContext(
-                p_persistence=self._persistence) as session_context:
+        with SessionContext(p_persistence=self._persistence) as session_context:
             request = flask.request
             with tools.TimingContext(lambda duration: self.measure(p_hostname=request.remote_addr,
                                                                    p_service=self.simplify_url(request.url_rule),
@@ -698,8 +692,7 @@ class StatusServer(base_web_server.BaseWebServer):
         with tools.TimingContext(lambda duration: self.measure(p_hostname=request.remote_addr,
                                                                p_service=self.simplify_url(request.url_rule),
                                                                p_duration=duration)):
-            with little_brother.persistence.session_context.SessionContext(
-                    p_persistence=self._persistence) as session_context:
+            with SessionContext(p_persistence=self._persistence) as session_context:
                 try:
                     user_infos = self._appcontrol.get_user_status_infos(p_session_context=session_context)
                     page = flask.render_template(
@@ -721,8 +714,7 @@ class StatusServer(base_web_server.BaseWebServer):
     @BLUEPRINT_ADAPTER.route_method("/about", endpoint="about")
     def about_view(self):
 
-        with little_brother.persistence.session_context.SessionContext(
-                p_persistence=self._persistence) as session_context:
+        with SessionContext(p_persistence=self._persistence) as session_context:
             request = flask.request
 
             with tools.TimingContext(lambda duration: self.measure(p_hostname=request.remote_addr,