+++ little_brother/test/test_status_server_devices.py
@@ -29,6 +29,12 @@ from little_brother.persistence.session_context import SessionContext
 from little_brother.test.base_test_status_server import BaseTestStatusServer
 from python_base_app.test import base_test
 
+NEW_DEVICE_NAME = "Some Device"
+NEW_DEVICE_HOST_NAME = "127.0.0.1"
+NEW_DEVICE_MIN_ACTIVITY_DURATION = 321
+NEW_DEVICE_MAX_ACTIVE_PING_DELAY = 21
+NEW_DEVICE_SAMPLE_SIZE = 32
+
 
 class TestStatusServerDevices(BaseTestStatusServer):
 
@@ -80,13 +86,58 @@ class TestStatusServerDevices(BaseTestStatusServer):
         delete_button.click()
 
         delete_button = self._driver.find_element_by_id("delete_device_1-modal-confirm")
-        # See https://stackoverflow.com/questions/56194094/how-to-fix-this-issue-element-not-interactable-selenium-python
-        self._driver.execute_script("arguments[0].click();", delete_button)
+
+        self.click(delete_button)
 
         with SessionContext(self._persistence) as session_context:
             device = device_entity_manager.get_by_id(
                 p_session_context=session_context, p_id=device_id)
             self.assertIsNone(device)
+
+    @base_test.skip_if_env("NO_SELENIUM_TESTS")
+    def test_page_devices_edit_device(self):
+        self.create_dummy_status_server()
+        self.create_selenium_driver()
+
+        self._status_server.start_server()
+
+        device_entity_manager: DeviceEntityManager = dependency_injection.container[DeviceEntityManager]
+
+        with SessionContext(self._persistence) as session_context:
+            device_id = device_entity_manager.add_new_device(
+                p_session_context=session_context, p_name_pattern=constants.DEFAULT_DEVICE_NEW_NAME_PATTERN)
+
+        self.login_devices()
+
+        elem_prefix = "device_{id}_".format(id=device_id)
+
+        elem = self._driver.find_element_by_id(elem_prefix + "device_name")
+        self.set_value(p_elem=elem, p_value=NEW_DEVICE_NAME)
+
+        elem = self._driver.find_element_by_id(elem_prefix + "hostname")
+        self.set_value(p_elem=elem, p_value=NEW_DEVICE_HOST_NAME)
+
+        elem = self._driver.find_element_by_id(elem_prefix + "min_activity_duration")
+        self.set_value(p_elem=elem, p_value=NEW_DEVICE_MIN_ACTIVITY_DURATION)
+
+        elem = self._driver.find_element_by_id(elem_prefix + "max_active_ping_delay")
+        self.set_value(p_elem=elem, p_value=NEW_DEVICE_MAX_ACTIVE_PING_DELAY)
+
+        elem = self._driver.find_element_by_id(elem_prefix + "sample_size")
+        self.set_value(p_elem=elem, p_value=NEW_DEVICE_SAMPLE_SIZE)
+
+        save_button = self._driver.find_element_by_id("save")
+        self.click(save_button)
+
+        with SessionContext(self._persistence) as session_context:
+            device: Device = device_entity_manager.get_by_id(
+                p_session_context=session_context, p_id=device_id)
+            self.assertIsNotNone(device)
+            self.assertEqual(NEW_DEVICE_NAME, device.device_name)
+            self.assertEqual(NEW_DEVICE_HOST_NAME, device.hostname)
+            self.assertEqual(NEW_DEVICE_MIN_ACTIVITY_DURATION, device.min_activity_duration)
+            self.assertEqual(NEW_DEVICE_MAX_ACTIVE_PING_DELAY, device.max_active_ping_delay)
+            self.assertEqual(NEW_DEVICE_SAMPLE_SIZE, device.sample_size)
 
 
 if __name__ == "__main__":